55 update renderer to load lod from generated geotiff files (#67)
* #55: Refined README

* #55: Added AnariHeightField class

* #55: Skip buggy property from visgl

* #55: Added AnariInstanceSet

Allows to manage instance array of AnariWorld so that children of
AnariWorld (e.g. AnariHeightField) can dynamically add/remove instances
from the world. This will be useful for loading/unloading DEM tiles.

* #55: Added updateAsync() and integrated to frame pipeline

The new function allows the world to update itself asynchronously
while we wait for the current frame to be complete and copy it over to
our output image.

The AnariHeightField class now has an empty updateAsync() function
called by AnariWorld.

* #55: Created AnariHeightFieldTile

* #55: Implemented basic multiple tile loading

* #55: Added LOD UI control attached to height field

* #55: Added support of tile mask

* #55: Implemented ANARI debug mode

* Create a debug device when debug mode enabled
* Fixed warnings that appeared after enabling debug mode including:
- Use of "fovY" instead of "fovy" in camera
- Do not commit renderer if no changes occurred
- Do not commit world if no changes occurred

* #55: Improve handling of partial tiles

* #55: Added ability to rebuild tile pyramid using im3e_geo_processor